\name{Kang18_8vs8}
\title{\code{Kang18_8vs8} dataset}
\alias{Kang18}
\alias{Kang18_8vs8}

\description{
10x droplet-based scRNA-seq PBMC data from 8 Lupus patients 
before and after 6h-treatment with INF-beta.

The dataset contains 35635 features across 29065 cells, 
and no filtering or preprocessing has been applied.

The original data is deposited in the Gene Expression Ombnibus (GEO) 
under accession number GSE96583.
}

\details{
\itemize{
    \item{link to reference:}{
        https://www.ncbi.nlm.nih.gov/pubmed/29227470}
    \item{link to raw data:}{
        https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE96583}
}
}

\references{\insertRef{Kang2018}{muscData}}