\name{Sala19_2vs2}
\title{\code{Sala19_2vs2} dataset}
\alias{Sala19}
\alias{Sala19_2vs2}

\description{
10x scRNA-seq data from E8.5 mouse chimeric embryos in which Tal1-knockout
(KO) embryonic stem cells were injected into a wild-type (WT) blastocyst.

The dataset contains 29453  features across 56122 cells, 
and no filtering or preprocessing has been applied.

The original data is made available by the 
Cancer Research UK Cambridge Institute (see link below).
}

\details{
\itemize{
    \item{link to reference:}{
        https://www.ncbi.nlm.nih.gov/pubmed/30787436}
    \item{link to raw data:}{
        https://content.cruk.cam.ac.uk/jmlab/chimera_tal1_data}
}
}

\references{\insertRef{Sala2019}{muscData}}